% Generated by roxygen2 (4.0.1): do not edit by hand
\name{rjitter}
\alias{rjitter}
\title{restricted jitter}
\usage{
rjitter(pts, polys, radius, maxiterations = 100)
}
\arguments{
\item{pts}{a SpatialPointsDataFrame of locations}

\item{polys}{a SpatialPolygonsDataFrame of regions}

\item{radius}{the radius of the disk}

\item{maxiterations}{limit of iterations}
}
\value{
A SpatialPolygonsDataFrame like pts but with new coordinates
}
\description{
jitter points but keep within regions
}
\details{
jitter points in a uniform disk but keep them within a region
}
\author{
Barry Rowlingson
}

